refactor: move paginator to helper collection
