+++ src/website/webview/paginators.py
@@ -20,3 +20,21 @@ class CustomCountPaginator(Paginator):
     @cached_property
     def count(self) -> int:  # type: ignore[override]
         return self.custom_count()
+
+class LargeTablePaginator(Paginator):
+    """
+    Overrides the count method to get an estimate instead of actual count when not filtered
+    """
+
+    _count = None
+
+    @property
+    def count(self) -> int:
+        """
+        Changed to use an estimate if the estimate is greater than 10,000
+        Returns the total number of objects, across all pages.
+        """
+        if self._count is None:
+            self._count = self.object_list.model.objects.count()
+
+        return self._count
